+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Panel - Define the hover area to be the same size of font size. (HaullyGames)
 * RemoteTech - Added signal strength based on Dist/maxDist. (HaullyGames)
 * Increased SCANSat experiment sizes. A planetary map shouldn't be smaller than a temperature reading. (Sir Mortimer)
+* Connection Manager Panel, click in signal Icon to open the Connection Panel. (HaullyGames)
 
 ------------------------------------------------------------------------------------------------------